adk/pyproject.toml
[project]
name = "agentic-layer-sdk-adk"
version = "0.1.0"
dependencies = [
    "google-adk[a2a]",
    "litellm",
    "opentelemetry-exporter-otlp-proto-http",
    "opentelemetry-instrumentation-starlette",
    "openinference-instrumentation-google-adk",
    "opentelemetry-instrumentation-httpx",
]

[dependency-groups]
test = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agenticlayer"]
